Changing to displacement-only optimization
It is not actually necessary to do a load-based simulation to get preload behavior in this case. All that is required is the stiffness, which should be linear for well beyond the displacement distance of choice. For better computation time, the code was adjusted to only use a two-part displacement simulation. Logging and a few other small changes were made as well.