Added checks to ensure stl normals face outward
Errors can happen when the displacement.pmx file is based on stl files with normals facing a certain way. If the opposite is ever used as a replacement, then whatever actions are inherent to the .pmx file will also be applied, inappropriately, to the replacement model. This causes meshing errors. A normal check and forcing action was added to ensure all models that are fed through the optimization will avoid this error.